create noise module, move dropout to it, add Gaussian noise, update docs
