+++ docs/sources/models.md
@@ -38,7 +38,8 @@ __Examples__:
 
 ```python
 from keras.models import Sequential
-from keras.layers.core import Dense, Dropout, Activation
+from keras.layers.core import Dense, Activation
+from keras.layers.noise import Dropout
 from keras.optimizers import SGD
 
 model = Sequential()
