+++ keras/layers/core.py
@@ -45,29 +45,6 @@ class Layer(object):
         return {"name":self.__class__.__name__}
 
 
-class Dropout(Layer):
-    '''
-        Hinton's dropout.
-    '''
-    def __init__(self, p):
-        super(Dropout,self).__init__()
-        self.p = p
-
-    def output(self, train):
-        X = self.get_input(train)
-        if self.p > 0.:
-            retain_prob = 1. - self.p
-            if train:
-                X *= srng.binomial(X.shape, p=retain_prob, dtype=theano.config.floatX)
-            else:
-                X *= retain_prob
-        return X
-
-    def get_config(self):
-        return {"name":self.__class__.__name__,
-            "p":self.p}
-
-
 class Activation(Layer):
     '''
         Apply an activation function to an output.
